Upgrade flatbuffers for all langs (#242)

+++ src/java/src/main/java/org/wololo/flatgeobuf/GeometryOffsets.java
@@ -2,7 +2,7 @@ package org.wololo.flatgeobuf;
 
 public class GeometryOffsets {
     public int coordsOffset;
-    public int[] ends = null;
+    public long[] ends = null;
     public int[] lengths = null;
     public int endsOffset = 0;
     public int lengthsOffset = 0;
